The Settings are Dead. Long live the Settings.
