+++ .idea/artifacts/Toolcat_jvm_1_0_1.xml
@@ -1,0 +1,8 @@
+<component name="ArtifactManager">
+  <artifact type="jar" name="Toolcat-jvm-1.0.1">
+    <output-path>$PROJECT_DIR$/build/libs</output-path>
+    <root id="archive" name="Toolcat-jvm-1.0.1.jar">
+      <element id="module-output" name="Toolcat.jvmMain" />
+    </root>
+  </artifact>
+</component>
