+++ src/jvmMain/kotlin/dev/anthonyhfm/toolcat/core/platform/DevicePlatformRepository.kt
@@ -1,0 +1,9 @@
+package dev.anthonyhfm.toolcat.core.platform
+
+interface DeviceRepository<Real, Simulator> {
+    var connectedDeviceList: MutableList<Real>
+    var simulatedDeviceList: MutableList<Simulator>
+
+    fun fetchConnectedDevices()
+    fun fetchSimulatedDevices()
+}
