+++ src/jvmMain/kotlin/dev/anthonyhfm/toolcat/main/ToolcatWindow.kt
@@ -4,11 +4,12 @@ import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.window.Window
 import androidx.compose.ui.window.application
 import dev.anthonyhfm.toolcat.commandline.CommandRegistry
-import dev.anthonyhfm.toolcat.core.platform.android.AndroidDeviceRegistry
+import dev.anthonyhfm.toolcat.core.platform.android.AndroidDeviceRepository
+import dev.anthonyhfm.toolcat.core.utils.GlobalSettings
+import dev.anthonyhfm.toolcat.main.theme.ToolcatTheme
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
-import settings.GlobalSettings
 import ui.views.MainView
 
 fun main(args: Array<String>) = application {
@@ -21,11 +22,14 @@ fun main(args: Array<String>) = application {
     ) {
         val scope = rememberCoroutineScope()
 
-        MainView()
+        ToolcatTheme.runCompose {
+            // MainView()
+            ToolcatMainView()
+        }
 
         scope.launch(Dispatchers.IO) {
             while (true) {
-                AndroidDeviceRegistry.scanDevices()
+                AndroidDeviceRepository.fetchConnectedDevices()
 
                 delay(100)
             }
