+++ src/jvmMain/kotlin/dev/anthonyhfm/toolcat/main/theme/schemes/StandardColorScheme.kt
@@ -1,0 +1,25 @@
+package dev.anthonyhfm.toolcat.main.theme.schemes
+
+import androidx.compose.material3.darkColorScheme
+import androidx.compose.material3.lightColorScheme
+import androidx.compose.ui.graphics.Color
+import dev.anthonyhfm.toolcat.main.theme.ToolcatColorSchemeSet
+import dev.anthonyhfm.toolcat.main.theme.TitleBarColors
+
+class StandardColorScheme : ToolcatColorSchemeSet {
+    override val name: String = "Standard Theme"
+
+    override val lightColorScheme = lightColorScheme()
+    override val lightTitleBarColors = TitleBarColors(
+        background = Color(0xff6750a4),
+        text = Color.White,
+        icons = Color.White
+    )
+
+    override val darkColorScheme = darkColorScheme()
+    override val darkTitleBarColors = TitleBarColors(
+        background = Color(0xff34333a),
+        text = Color.White,
+        icons = Color.White
+    )
+}
