+++ src/jvmMain/kotlin/dev/anthonyhfm/toolcat/modules/settings/presets/ScaleSetting.kt
@@ -1,0 +1,17 @@
+package dev.anthonyhfm.toolcat.modules.settings.presets
+
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.MutableState
+
+class ScaleSetting(
+    val displayText: String,
+    val value: MutableState<Int>,
+    val minValue: Int,
+    val maxValue: Int,
+    val onValueChange: (Int) -> Unit
+) : SettingModel {
+    @Composable
+    override fun Setting() {
+
+    }
+}
