+++ src/jvmMain/kotlin/dev/anthonyhfm/toolcat/modules/settings/presets/SettingModel.kt
@@ -1,0 +1,8 @@
+package dev.anthonyhfm.toolcat.modules.settings.presets
+
+import androidx.compose.runtime.Composable
+
+interface SettingModel {
+    @Composable
+    fun Setting()
+}
