--- src/jvmMain/kotlin/settings/SettingsCluster.kt
@@ -1,6 +1,0 @@
-package settings
-
-data class SettingsCluster(
-    val title: String,
-    val settings: Array<SettingsViewModel>
-)
