--- src/jvmMain/kotlin/settings/views/CheckForUpdatesSettingsModel.kt
@@ -1,48 +1,0 @@
-package settings.views
-
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.height
-import androidx.compose.material3.Button
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.runtime.*
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import settings.SettingsViewModel
-
-class CheckForUpdatesSettingsModel : SettingsViewModel {
-    @Composable
-    override fun content() {
-        Row(
-            modifier = Modifier
-                .fillMaxSize()
-                .height(40.dp),
-
-            horizontalArrangement = Arrangement.SpaceBetween,
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Row(
-                verticalAlignment = Alignment.CenterVertically,
-                horizontalArrangement = Arrangement.spacedBy(16.dp)
-            ) {
-                Text(
-                    text = "Software Update",
-                    fontSize = 18.sp,
-                    color = MaterialTheme.colorScheme.onBackground
-                )
-            }
-
-            Button(
-                onClick = {
-                    TODO("Check for any new Github Releases")
-                }
-            ) {
-                Text("Check for Updates")
-            }
-        }
-    }
-}
