--- src/jvmMain/kotlin/settings/views/CustomWindowTitleBarModel.kt
@@ -1,56 +1,0 @@
-package settings.views
-
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.height
-import androidx.compose.material.Text
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Switch
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import settings.GlobalSettings
-import settings.SettingsViewModel
-
-class CustomWindowTitleBarModel : SettingsViewModel {
-    @Composable
-    override fun content() {
-        val checked by remember { mutableStateOf(GlobalSettings.customTitleBar) }
-
-        Row(
-            modifier = Modifier
-                .fillMaxSize()
-                .height(40.dp),
-
-            horizontalArrangement = Arrangement.SpaceBetween,
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Row(
-                verticalAlignment = Alignment.CenterVertically,
-                horizontalArrangement = Arrangement.spacedBy(16.dp)
-            ) {
-                Text(
-                    text = "Enable the Custom Title Bar",
-                    fontSize = 18.sp,
-                    color = MaterialTheme.colorScheme.onBackground
-                )
-            }
-
-            Switch(
-                checked = checked.value,
-                onCheckedChange = {
-                    checked.value = it
-
-                    GlobalSettings.customTitleBar = checked
-                    GlobalSettings.saveGlobalSettings()
-                }
-            )
-        }
-    }
-}
