--- src/jvmMain/kotlin/ui/Font.kt
@@ -1,44 +1,0 @@
-package ui
-
-import androidx.compose.ui.text.font.FontFamily
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.platform.Font
-
-val Inter = FontFamily(
-    Font(
-        resource = "fonts/Inter-Thin.ttf",
-        weight = FontWeight.Thin
-    ),
-    Font(
-        resource = "fonts/Inter-ExtraLight.ttf",
-        weight = FontWeight.ExtraLight
-    ),
-    Font(
-        resource = "fonts/Inter-Light.ttf",
-        weight = FontWeight.Light
-    ),
-    Font(
-        resource = "fonts/Inter-Medium.ttf",
-        weight = FontWeight.Medium
-    ),
-    Font(
-        resource = "fonts/Inter-Regular.ttf",
-        weight = FontWeight.Normal
-    ),
-    Font(
-        resource = "fonts/Inter-SemiBold.ttf",
-        weight = FontWeight.SemiBold
-    ),
-    Font(
-        resource = "fonts/Inter-Bold.ttf",
-        weight = FontWeight.Bold
-    ),
-    Font(
-        resource = "fonts/Inter-ExtraBold.ttf",
-        weight = FontWeight.ExtraBold
-    ),
-    Font(
-        resource = "fonts/Inter-Black.ttf",
-        weight = FontWeight.Black
-    ),
-)
