--- src/jvmMain/kotlin/ui/theme/ToolcatColorScheme.kt
@@ -1,18 +1,0 @@
-package ui.theme
-
-import androidx.compose.material3.ColorScheme
-import androidx.compose.ui.graphics.Color
-
-data class TitleBarColors(
-    val background: Color,
-    val text: Color,
-    val icons: Color
-)
-
-interface ToolcatColorSchemeSet {
-    val name: String
-    val lightColorScheme: ColorScheme
-    val lightTitleBarColors: TitleBarColors
-    val darkColorScheme: ColorScheme
-    val darkTitleBarColors: TitleBarColors
-}
