+++ src/jvmMain/kotlin/ui/views/MainView.kt
@@ -35,25 +35,11 @@ fun MainView() {
             icon = { Icon(painterResource("icons/cast.svg"), "") },
             view = { CastView() }
         ),
-        NavigationItem(
-            name = "Settings",
-            icon = { Icon(Icons.Default.Settings, "") },
-            view = { SettingsView() }
-        ),
-        NavigationItem(
-            name = "About",
-            icon = { Icon(Icons.Default.Info, "") },
-            view = { AboutView() }
-        )
     )
 
     Row {
         NavigationRail {
             navigationItems.forEachIndexed { index, item ->
-                if (index == navigationItems.count() -1) {
-                    Spacer(modifier = Modifier.weight(1F))
-                }
-
                 NavigationRailItem(
                     icon = { item.icon() },
                     label = { Text(item.name) },
